requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptg==0.4.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
duplicity==0.8.21
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openai==0.26.5
openapi==1.1.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.4.0
protobuf==3.12.4
proton-client==0.7.1
protonvpn-cli==3.13.0
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.9.0
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.6
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-discord==1.7.3
python-dotenv==1.0.0
python-gnupg==0.4.8
pythondialog==3.5.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==4.5.1
reportlab==3.6.8
requests==2.28.2
requests-toolbelt==0.9.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
spotipy==2.22.1
systemd-python==234
Telethon==1.27.0
termcolor==2.2.0
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.8.2
zipp==1.0.0
